#217
The arrangement of the element look weird, especially what's happening below the title: to much spacing. Might be better to use the format that's used in all the pages: a top and a bottom? Top part should have the paddings it usually has (see every other page in New account). The white warning-box should just be at the top of the bottompart, with a margin to the icon-button (mobile: 50px, desktop: 70px)